requirements.txt
torch torchvision --index-url https://download.pytorch.org/whl/cu121
ultralytics==8.2.0
timm==1.0.9
pytorch-metric-learning==2.5.0
scikit-learn==1.4.2
onnx==1.16.2
onnxruntime-gpu==1.18.0
faiss-gpu==1.8.0
opencv-python==4.10.0.84
Pillow==10.3.0
numpy==1.26.4
tqdm==4.66.4
